Correlation: Fix bug with international characters
JIRA:MADLIB-1186

Additional Author: Nandish Jayaram <njayaram@apache.org>

If the column name of an independent variable used in
madlib.correlation(...) has quotes in it, then the query fails due
to a regular string concat used for creating an intermediate column
name that reflects the average of the column.

This commit uses add_postfix() to create that column name instead.
Originally, the new column name was `avg_{column_name}`, that is
replaced with add_postfix(column_name, '_avg'). The prefix `avg_`
is changed to suffix `_avg`. This is only an intermediate column,
and not shown as an output, hence ignoring the semantics of the
final string name.

Closes #214