use gd instead of sd
like sd, this code also assumes that fit and evaluate run in the same
process so the gd can be shared across function calls

fit:
since we use gd, we can reuse the same session for all calls to fit and
evaluate. This session will get reset at the final call to evaluate

Only the very final evaluate should clear the
session which may not necessarily be the training evaluate. changed the
code to take care of this scenario

fit_multiple:
we need to clear gd and also the session for each hop for each call to
fit/evaluate step functions. for fit multiple, the session should be
cleared at the end of each hop i.e. last row of each call to the
uda/step function

todo:
1. what if evaluate is called directly ?? need to test
2. test fit/fit multiple with gpu