Send Traceback of stack from segment back to coordinator
This makes it much easier to debug when something in an inner
UDA or UDF running on a segment causes an Exception, when
it's called from an outer UDF running on master.  Usually,
you only get to see the stack trace of the outer UDF, which
is not where the bug is, so the only thing you learn is that
it's somewhere inside the UDA or UDF it's calling--not very
useful.  With these changes, the full stack trace gets fowarded
back to the coordinator and added to the Exception detail there,
so you get to see both, including the line number where the
Exception occurred.