DL: Fix error messages in fit multiple
JIRA: MADLIB-1464

Previously while calling fit_multiple, passing NULL to the mst_table
argument would print the wrong error message and passing NULL to the
output table did print the right error message but the message wasn't
consistent with other modules. This commit fixes both of these issues.

Refactored the validator code such that we don't need to create the info
and summary table names in the fit multiple class. Instead we do that in
the validator and then the validator object can be used to get the table
names. This makes it easier to validate all the tables inside the
validator class.

To test this, we created two plpy functions that would make sure that we
print the right error message for invalid input and output tables.

Co-authored-by: Ekta Khanna <ekhanna@vmware.com>